requirements.txt
pandas>=1.5.0
numpy>=1.24.0
requests>=2.31.0
streamlit>=1.25.0
plotly>=5.15.0
folium>=0.14.0
geopandas>=0.13.0
shapely>=2.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.4.0
